requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
sqlalchemy==2.0.35
jinja2==3.1.4
python-multipart==0.0.12
python-jose[cryptography]==3.3.0
passlib==1.7.4
openpyxl==3.1.5
aiofiles==24.1.0
psycopg2-binary==2.9.9
